remove uses of "npm bin"
This was removed from recent Node versions. Using 'npx' is the new way
and should resolve this issue.

I suppose we might run into someone who had a non-working npx and it
breaks, but we'll cross that bridge when we come to it; it works in my
testing and there are people broken in the current version.