Add "pull" command for updating from the remote
The theory here is much like "commit": most of the time, you don't care
about Git minutiae with your Zettelkasten and just want to be able to
run a command that Makes Things Work™. "pull" thus checks if there are
upstream changes, and if there are stashes anything unsaved in the
local, rebases, and if needed pops the stash back off.

If there are merge conflicts this will crash, but this seems like an
extremely unusual case for most workflows (I've almost never encountered
them). It's also easy to fix: just resolve the conflicts using normal
Git operations and, if necessary, pop your stash.